// src/app/models/user.ts
export interface User {
    userId: number;
    firstName:string
    lastName: string;
    emailId: string;
    mobileNumber: number;
    password: string;
    nationality: string;
    passportNumber: number;
    permanentAddress:string;
    officeAddress:string;
    personalIdentificationNumber:string
  }